mega commit

+++ PRPA/Models/Service.cs
@@ -5,7 +5,7 @@
         public int ServiceId { get; set; }
         public string Name { get; set; }
         public int Duration { get; set; }
-        public float Cost { get; set; }
+        public decimal Cost { get; set; }
         public string Description { get; set; }
     }
 }
